Move SlotQueue to internal module for better code organization
- Move SlotQueue, _PutSlot, and _GetSlot classes from kioto/streams/impl.py to kioto/internal/buffer.py
- Update imports in streams implementation to use internal module
- Consolidate internal infrastructure components in single module
- Maintain all existing functionality and API compatibility
- All 126 tests continue to pass with no regressions
